Update GEOS-Chem timers for better readability
Updated the print statements for the GEOS-Chem timers so that the results
show in a formatted table, making the output easier to read and compare.
We now print the timer in the DD-hh-mm-ss.SSS format.

Also changed many of the GEOS-Chem timer functions to subroutines since they
do not actually return a value, except for RC which we usually pass as an
INTENT(INOUT) argument. Also bracketed all calls to GEOS_Timer_* so that
we do not enter these routines unless USE_TIMERS is defined.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>