Update diag49_mod.F to print out tracer mass if using the passive tracer
Lizzie Lundgren had already updated diag49_mod.F to print out tracer mass
to a text file if compiling with MASSCONS=yes. We wanted to use the same
method for printing out the instantaneous passive tracer mass. To do this,
we removed CPP statements checking if MASSCONS is defined and instead use
IF statements to check if the logical DO_MASSCONS is true or if IDTPASV>0.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>