Restore fast "clean" in the Makefiles
In v10-01 we had rewritten the "clean" targets to the Makefiles
to only remove the files that were created in each subfolder.  This
was necessary to do a "make realclean except RRTMG".  However, the
new algorithm was much slower than the old one.

We have now restored the faster algorithm for the "clean" targets
(i.e. rm -f *.o *.mod *.a. *.x).  We also have preserved the older
algorithm under the "slowclean" target name.

Lastly, we have added a new Makefile target "realclean_except_rrtmg",
which will use the "slowclean" targets in the various makefiles.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>