Now assign actual MW (e.g. MW_g) values to XMW in drydep_mod.F
Historically, routine INIT_DRYDEP used a very large IF-ELSE statement
to define species-specific properties for dry deposition.  This has
largely been removed by the implementation of the species database in
v11-01d.

But we had added some kludges to v11-01d so that we could preserve
the original behavior of prior versions.  For the most part, this meant
leaving incorrect molecular weight definitions.

We now use the actual molecular weight (MW_g) from the species database
to assign the XMW array in drydep_mod.F.  This will result the following
changes (units are g/mol):

                new XMW   old XMW
  ACET:           58.08      58.0
  ALD2:           44.05      44.0
  ALK4:           58.12      12.0  ==> formerly used emitted MW
  BC & OC:        12.01      12.0
  OPOA & OPOG:    12.01      12.0
  MTPO           136.23     136.0
  NITs            62.0       36.0  ==> formerly used sea salt MW?
  SO4s            96.0       36.0  ==> formerly used sea salt MW?
  SALA            31.4       36.0  ==> update from v9-02
  SALC            31.4       36.0  ==> update from v9-02

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>