Bug fix to update air quantities in DO_PBL_MIX2 after humidity update.
Specific humidity is updated in VDIFFDR. Since many air quantities are
now dependent on specific humidity, air quantities should be updated
following the moisture update by calling AIRQNT.

Since units are in v/v in mixing, units are converted to kg prior to the
call to AIRQNT, and back to v/v after. This is necessary to conserve
tracer mass as the air quantities change.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>