Give SO4s and NITs the same molecular weight as SALC = 31.4 g/mol
We have now resolved an inconsistency in the molecular weights of
of the SO4s and NITs species (i.e. SO4 and NIT on coarse sea salt
aerosol).

In the input.geos files:
  SO4s had a MW of 96 and
  NITs had a MW of 62.

In drydep_mod.F:
  XMW = 36 for SO4s
  XMW = 36 for NITs

It has been determined (see comments in code) that because SO4s and
NITs are both internally mixed into coarse sea salt, it is appropriate
to use the molecular weight of coarse sea salt (now = 31.4 g/mol) for
both SO4s and NITs.

We updated the species database definition of SO4s and NITs to use
a molecular weight of 31.4 g/mol.  The same modification will also
be made to the input.geos files.

%%% NOTE: in a future commit, we will remove the molecular weights
%%% from the input.geos file, and use only the values from the
%%% GEOS-Chem species database.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>