Make sure units of State_Chm%Species are always molec/cm3
In the new FlexChem code, the species concentration array
(State_Chm%Species) was initialized in v/v in INIT_FLEXCHEM
and converted to molec/cm3 in FLEX_CHEMDR just prior to calling
the chemistry solver. However, the previous practice was to
always store State_Chm%Species in molec/cm3. To ensure
consistency throughout the code, we now initialize the species
in v/v and immediately convert to molec/cm3 in INIT_FLEXCHEM.
This way, the unit conversion for the species array is no
longer needed in FLEX_CHEMDR.

A bug in the STTTOCSPEC array has also been fixed. In this
array, the STT tracers were being mapped to the KPP species
when they should have been mapped to the GEOS-Chem species.

Some of the debug print statements have also been beefed up.
These can be used to track OH and NO concentrations in
State_Chm%Species. The debug code can be removed once we're
satisfied with the FlexChem results.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>