Minor fixes to several ProTeX headers
Added a backslash before several underscores that occur in
the DESCRIPTION section of the ProTeX headers. Also updated
doc/Makefile_SrcDoc.mk for the recent updates to the chemistry
routines. Finally, ensured that reference documentation is now
created without error when we type "make doc" in the top-level
code directoy.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>