Temporary fix for calculating air density
As a temporary fix, restore old method for calculating air
density (from RURALBOX) to remove any differences caused by the
unit conversions. The AIRDEN calculation has been moved to
after AD is calculated since it uses that field. Once we are
satisfied with the FlexChem results, we restore the AIRNUMDEN
calculation using the ideal gas law.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>