Hydrophobic aerosol properties (fix 2 of 2)
This fix is from Seb Eastham:

When calculating total SAD and radius of BC and OC, the properties of
the hydrophobic component were calculated using a radius which was
given in um but expected to be in cm. As a result, the additional SAD
due to hydrophobic aerosol was underestimated by x10,000, and the
average aerosol radius was overestimated in a similar fashion. This
has been corrected by applying a 1.0D-4 factor wherever RW(1) is
referenced. This is the second of two connected corrections.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>