Rename SET_BUFFER_SIZE to LINE_BUFFER in file_mod.F
We have renamed the SET_BUFFER_SIZE routine in GeosUtil/file_mod.F to
LINE_BUFFER.  This will make the subroutine easier to understand.

LINE_BUFFER now only takes one argument, the Fortran logical unit
number.  It calls SETVBUF3F( UNIT, 1, 80 ), which sets the file to
line-buffering with a buffer size of 80 chars.  This should be suffcient
for logfile output.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>