Initial compiler flag settings for PGI compiler in Makefile_header.mk
In Makefile_header.mk:

(1) The default value of Makefile variable $(COMPILER) now is taken
    from the $(FC) environment variable.  This is set in the .bashrc
    or when a module is loaded.

(2) Make the test for the PGI fortran compilers more robust (i.e.
    it should work for "pgf77", "pgf90", or "pgfortran".)

(3) Add -Kieee to specify IEEE floating-point math.

(4) Add -Ktrap=fp for the FPE=yes option.  This should be analogous
    to the ifort "-fpe0 -ftrapuv" setting.

NOTE: The compiler flags may still change.  We will determine the
optimal compiler settings so that we will obtain similar output as
compared to when we compile with ifort.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>