Fix unit conversion of STT array in flex_chemdr.F
We need to keep STT in units kg when we do chemistry because
tracer concentrations from STT are used in gckpp_HetRates.F90.
This is to reproduce code that was formerly in calcrate.F.

When we remove family tracers, we can get rid of the multiple
unit conversions in flex_chemdr.F and do the unit conversions
between kg and molec/cm3 in the same step that we copy between
STT and State_Chm%Tracers. That way, the STT array will remain
in units kg and avoid confusion.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>