Implement met-field netcdf diagnostics for GEOS-FP and MERRA2.
In order to enable hourly diagnostics starting with the first hour of
a simulation, MET field netcdf diagnostics are updated via a call to
DiagnUpdate_Met in diag1.F. This differs from bpch which updates dao
fields when they are read in. This approach is not possible with netcdf
diagnostics using HEMCO since the met-fields may be read in intervals
greater than 1 hr (e.g. every 3 or 6 hrs). For in between hours, the
diagnostics would not be updated or output. This would not affect daily
mean diagnostics but would prevent hourly diagnostic output.

This update currently only enables netcdf diagnostics for GEOS-FP and
MERRA2.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>