Add tagged Hg species to the species database
Modified the Hg0, Hg2, and HgP statements in species_database_mod.F90 to
include the tagged Hg tracers. Added the Hg_CAT and Hg_CAT_FULL arrays
to Input_Opt to include extra information needed to pass the proper Name
and FullName to routine Spc_Create. The Hg_CAT and Hg_CAT_FULL arrays are
filled in routine TRACERID, although this may be moved to mercury_mod.F in
the future as we work towards removing tracerid_mod.F as part of FlexChem.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>