Fix array-out-of-bounds error for tagged Hg simulation
In routine SRCHg0, the AD03_nat array was using N for the third index,
but N wasn't explicitly set for that purpose and so the last value of N
was used instead. This was causing array-out-of-bounds errors in ADO3_nat.
To fix this, we now loop over N = 1, N_Hg_CATS and archive ADO3_nat for
each tagged Hg category.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>