Apply fixes for HBr and HOBr rates to KPP/TempFlex/gckpp_HetRates.F90
Fixes were made in KPP/TempFlex/gckpp_HetRates.F90 for the HBr
and HOBr rates. As part of this fix, we moved the kludge from
Justin Parrella for setting the rates equal to one another from
function HETHBr to SET_HET. This was done so that the kludge
would apply to both HBr and HOBr rates.

Made other modifications so that gckpp_HetRates.F90 in the
KPP/TempFlex and KPP/TempFlex_Benchmark directories are
consistent with one another. This involved bracketing some
code specific to the UCX mechanism with CPP statements to avoid
executing it if we're using the tropchem mechanism.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>