Use BPCH_DIAG and NC_DIAG C pre-proc switches; hard-code species background
Species background values are stored in globchem.dat for use when the
CSPEC restart file is not available. For NetCDF, these values will be
assigned in read_gc_restart_nc if the NetCDF restart file does not
contain species data. The values are currently hard-coded there, but
eventually can be put into the species database for easier reference.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>