Move code for setting methanol concentrations to INIT_FLEXCHEM
We previously set MOH concentrations in flex_chemdr.F.
To be consistent with the rest of the species, we now set
these concentrations in INIT_FLEXCHEM (found in
GeosCore/flexchem_setup_mod.F90).

Combined the code for CSPEC restart not found and CSPEC restart
not being used since they did the same thing. Both situations
set species concentrations to the default values specified in
globchem.dat, so there's no sense in duplicating the code.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>