Block off KPP diagnostics with #if defined( DEVEL )
Only initialize and save out KPP diagnostics (rates and species
concentrations) to netCDF if the user compiles with DEVEL=y.

Also modified debug output so that we print out species
concentrations from State_Chm%Species instead of the C array.
This was done so that species concentrations are written to
the log file in the same order that they are in the reference
code.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>