Restore ND22 diagnostic in FlexChem
We used to update the ND22 diagnostic in calcrate.F, but that
routine is now obsolete. To restore ND22, we now update the
ND22 diagnostic in routine PHOTO_JX in fast_jx_mod.F. This way
the the ND22 diagnostic is updated immediately after we
calculate the J-values in the ZPJ array.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>