Bug fix to re-enable stratospheric emissions for specialty simulations
Hongyu Liu noticed a low Be7 stratospheric source and imbalanced
stratospheric + tropospheric sources and sinks observed in the v11-01f
1-yr RnPbBe benchmark. We traced the change to one line in HEMCO v1.1.016
that was brought into GEOS-Chem v11-01e which restricted adding emissions
above the chemistry grid to the UCX simulation only. This was applied
in routine DO_TEND within mixing_mod.F90:

   +          ! For non-UCX runs, never emit above the chemistry grid.
   +          ! (ckeller, 6/18/15)
   +          IF ( .NOT. Input_Opt%LUCX ) ChemGridOnly = .TRUE.

After removing this line and rerunning the 1-yr RnPbBe benchmark, the
Be7 values are on par with v11-01d benchmark results and the total
sources and sinks are balanced. We therefore adjusted the line to apply
only when using a fullchem simulation and UCX=false so that specialty
simulations are excluded.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>