Remove reaction hardwiring in PHOTORATE_ADJ routine
The routine PHOTORATE_ADJ (in GeosCore/fast_jx_mod.F) adjusts certain
photolysis reactions for FlexChem.  The reaction numbers used to be
hardwired to the same values in the FJX_j2j.dat file.

We have now introduced several reaction flags as module variables
to GeosCore/fast_jx_mod.F.  When we read the FJX_j2j.dat file in
routine RD_JS_JX, we scan through the list of reactants and products
to flag the reactions that will be touched by PHOTORATE_ADJ.

Bob Y. validated this modification with a 6-hour difference test
for the tropchem simulation.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>