Bug fix: uncomment copying of C to State_Chm%Species
In the prior commits, we inadvertently commented out the block of
code where the updated concentrations from KPP get copied back
into State_Chm%Species.  We have now restored this.

Also, we have blocked out the diagnostics for AVERAGEMASS and MASS
in an #ifdef DEVEL block, so that these can be printed out if desired.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>