Restore parallelization fixes for the benchmark mechanism
Some modifications made earlier for running FlexChem with
parallelization turned on were inadvertantly overwritten in
a later commit. The missing threadprivate statements are now
restored in gckpp_Function.F90, gckpp_Global.F90, and
gckpp_LinearAlgebra.F90 in the TempFlex_Benchmark directory.
We also need to comment out the EQUIVALENCE statements in
TempFlex_Benchmark/gckpp_Global.F90 again to get the benchmark
mechanism to work with parallelization turned on.

Also updated TempFlex/gckpp_Global.F90 so that the declaration
of array HET is consistent with the benchmark mechanism. We
should also update the gckpp.kpp file so that the inlined
code is consistent for any other mechanism that we build.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>