Turn off emission diagnostics if LEMIS is false
To prevent simulations from crashing with HEMCO errors when emissions are
turned off, we now include a check for LEMIS in READ_DIAGNOSTIC_MENU
(found in GeosCore/input_mod.F). If LEMIS is false, then we set NDxx for
the emission diagnostics to zero. We also print a warning to the log file
to notify the users that these diagnostics are being turned off.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>