Fix bug where tracer conc diags only created for NO in mp runs.
This issue is confined to FlexChem runs only, where tracer
concentration diagnostic containers are created in flexchem_setup_mod
rather than during diagnostic initialization in diagnostics_mod.
The ND45 diag initialization in diagnostics_mod has been and continues
to be commented out in FlexChem code.

The issue is due to passing an uninitialized cID to diagn_create
in flexchem_setup_mod. When using diagnostic container name as the
unique identifier for diagnostics, cID is not required. Passing the
same value results in failure to create a diagnostic, unless cID = -1.
For unknown reasons, cID was a very large integer for mp runs but -1
for sp runs and therefore only caused issues for mp runs.

This issue has been fixed by removing the cID argument from
diagn_create in flexchem_setup_mod.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>