Bug fix: remove AD03(I,J,21,1) update from ADD_Hg2_SNOWPACK
Routine ADD_HG2_SNOWPACK has a statement to update slot #21 of the
AD03 diagnostic, which was labeled as "Hg2/HgP deposition to snow/ice".

But it appears somewhere in a recent commit (maybe v10-01) that slot #21
of ADO3 is now "Production of Hg2 from rxn with O3".  Therefore the
line of code that updates ADO3 is no longer releveant.  We have now
commented this out.  Need to ask Team Hg for clarification.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>