Fix physical constants to be consistent with merged code
In v11-01f, GEOS-Chem gets the physical constants from physcontants.F
instead of comode_loop_mod.F. As such, code added for FlexChem needed to
be updated to get constants from that routine.

Also, the parameter CONSVAP is added to physconstants.F. This was previously
defined in comode_loop_mod.F. In the physical constants update for v11-01f,
the calculation for CONSVAP was moved to the two files (aerosol_mod.F and
planeflight_mod.F) where it was used and computed locally. However, in
FlexChem we utilize CONSVAP in a few more places (in aerosol_mod.F,
flexchem_setup_mod.F90, and gckpp_HetRates.F90). Therefore, this parameter
has been moved into physconstants.F so that it can be defined and computed
in one location.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>