Fixed a typo that was preventing ND29 (CO--SRCE) output
In GeosCore/diag3.F, we had added an IF block to set the NN value,
which is the # of slots of the ND29 diagnostic to print out.  The IF
statement was correct for the tagged CO simulation.

But for the full-chemistry simulations, the IF statement was setting
N = 6 instead of NN = 6.  This caused N to be undefined and to exit
the ND29 output DO loop to exit without iterating.  Now fixed.

Also removed a debug print statement from GeosCore/readchem.F.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>