Update Makefiles for new KPP directories
KPP_CHEM is now set accordingly:
   (1) KPP_CHEM=Standard  for the benchmark mechanism (72L, tropchem+SOA+UCX)
   (2) KPP_CHEM=SOA       for the standard SOA mechanism (47L)
   (3) KPP_CHEM=SOA_SVPOA for the SOA mechanism w/ semivolatile POA (47L)
   (4) KPP_CHEM=Tropchem  for the NOx-Ox-HC-Aerosol-Bromine mechanism (47L)
   (5) KPP_CHEM=UCX       for the UCX mechanism (72L)

The KPP/Makefile has also been updated for the new mechanism directories.

For now we will ignore the KPP/TempFlex* directories. These were added
during the original implementation of FlexChem, which was based on an
intermediate version of v11-01c. They are kept here for reference and can
be removed when FlexChem is fully functional in what will be v11-01g.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>