Restore archiving of AD03(I,J,21,1) -- slot #4 of PL-HG2-$
Per request of Jenny Fisher, we have restored the archiving of
AD03(I,J,21,1) in GeosCore/depo_mercury_mod.F.  The results of the
unit tests are "yellow", meaning the diagnostic outputs of runs
done with single-processor and multi-processors differ, but the
restart files are identical.

We think that a combination of writing to this 2-D diagnostic from a
3-D parallel loop, plus numerical roundoff when casting REAL*8 values
to REAL*4, may be the root cause of this issue.

Because only the diagnostics differ, we can consider the unit tests to
have "passed with warnings".  We probably will not be able to fix this
issue without drastically rewriting the program flow in the Hg code.
Therefore, leave it as-is for now, and just note that it happens.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>