Add fixes for KPP/Tropchem/gckpp_Rates.F90
The gckpp files were regenerated to fix an error in the reaction
GLYX + NO3 -> HNO3 + HO2 + 2CO. The function for that reaction should be
GC_GLYXNO3, not GCARR.

In addition, the previous gckpp files for Tropchem had been created with
the wrong gckpp.kpp file, so an unnecessary function for GC_CRIEGEE was
included and the GC_GLYXNO3 function was missing. The GC_CRIEGEE routine
is not needed because we can simply add H2O as a reactant to apply Criegee
chemistry.

Here we recreate the gckpp_Rates.F90 file with the proper gckpp.kpp file.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>