Fix for copying H2O to the C array in flex_chemdr.F for non-UCX simulations
The value of H2O was calculated, but never copied to the C array for KPP.
Several reactions require C(ind_H2O), which is copied to FIX(indf_H2O), to
compute the reaction rates in gckpp_Function.F90. Now, H2O is copied to
C if UCX is not defined. For UCX simulations, H2O is a tracer and
concentrations are obtained directly from the State_Chm%Species array.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>