Bug fix to include conditional for minute == 0 in ITS_A_NEW_MONTH().
The day 1 hour 0 condition in ITS_A_NEW_MONTH requires that GMT be less
than TS which can only occur if GMT has minute 00. This was always the
case when the chemistry timestep was 60 minutes since the next start time
after day 1 hour 0 minute 0 is day 1 hour 1. However, this is no longer the
case with a smaller chemistry timestep. As a result, starting a simulation
at minute 20 on day 1 hour 0 of the month resulted in an incorrect return
for ITS_A_NEW_MONTH since GMT was equal to TS. This caused seg faults in
simulations that use ITS_A_NEW_MONTH to read monthly data (specifically
Hg and tagged CO) since the data was not read in.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>