Add a placeholder KPP/Custom directory for the custom chemistry mechanism
The KPP/Custom directory is added. This directory will eventually be used
for building KPP on-the-fly during the GEOS-Chem build process. The idea
is that users will be able to pass CHEM=custom as a make option and
GEOS-Chem will build KPP and create the necessary gckpp*.F90 files. In this
manner, users will be able to modify the globchem.* files to create a
custom chemistry mechanism for GEOS-Chem.

The globchem.def, globchem.eqn, and globchem.spc files in KPP/Custom are
copied from the KPP/Standard directory. Makefile, gckpp.kpp, and
gckpp_HetRates.F90 are symbolic links that point to the files in
KPP/Standard. These can be made actual files in the future if we think
users might modify them and not intend for their changes to be propagated
to other mechanisms.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>