Bug fix: Don't set locally-defined pointers to NULL in declarations
A subtle parallelization error can occur when a pointer that is
used in an OpenMP parallel loop is nullified when it is declared.

Typing this declaration:

   TYPE(Species), POINTER :: ThisSpc => NULL()

will force ThisSpc to be a SAVEd pointer by default.  This can cause
problems if ThisSpc is used in an OpenMP loop, since loop pointers
should be created and nullified on each thread.

The simple solution is to nullify the pointer in the program, i.e.

   TYPE(Species), POINTER :: ThisSpc
   . . .
   ThisSpc => NULL()

NOTE: Any pointers that are module variables (i.e. located above the
CONTAINS statement) can be nullified when declared, since module variables
are globally SAVEd variables by default.

We need to go through the rest of the code to rewrite similar pointer
declarations.  This will be done in the following commits.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>