Continue removing local pointer nullifications from declarations
As with the previous commit, we have gone throughout GEOS-Chem and
have removed lines like

  REAL(fp), POINTER :: STT(:,:,:,:) => NULL()

with

  REAL(fp), POINTER :: STT(:,:,:,:)
  . . .
  STT => NULL()

This will avoid inadvertently setting the pointer to a SAVEd variable,
which may present problems in OpenMP parallel loops.

NOTE: This only applies to local pointers in subroutines.  Pointers that
are also module variables are not affected, since they are SAVEd anyway.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>