Remove IDTPOP* tracer ID flags from tracerid_mod.F
The FlexChem implementation necessitates the removal of tracerid_mod.F.
To facilitate this, we are removing the tracer ID flags for the specialty
simulations from tracerid_mod.F.  These are now defined as local module
variables where needed.

In this commit, we have removed all references to the IDTPOP* tracer ID
flags for the POPs simulation.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>