Add fixes for setting heterogeneous rates for PSC chemistry
In the HET*_PSC functions, move the IF statement for STRATBOX to be within
the loop over 1, NAERO and end the IF block after setting XSTKCF. This will
make sure that the code for computing ADJUSTEDRATE is executed even if a
grid box doesn't meet the STRATBOX criteria. This fix was made to reproduce
the code in calcrate.F.

Additionally, make sure to set XSTKCF = B if the grid box is not STRABOX.
Again, this reproduces code that was previously in calcrate.F. In the old
code, we set XSTKCF = BRR(NK,NCS) if NK was not equal to any of the NK*
flags for the various heterogeneous reactions or if not STRATBOX, where
BRR is the B rate coefficient originally read from globchem.dat.

Finally, added a minor fix to flex_chemdr.F to set ERR_MSG passed if
the integrator fails a second time.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>