Remove obsolete LKPP option
The LKPP option in the Chemistry Menu of input.geos is no longer needed.
With the introduction of FlexChem in v11-01g, SMVGEAR has been removed
and KPP is the only chemical solver available. As such, the LKPP option
is now obsolete.

Also added notes diag_pl_mod.F and READ_PROD_LOSS_MENU in input_mod.F
to remind us that the prod/loss diagnostics currently do not work in
v11-01g. The prod/loss diagnostics will be restored when we complete the
work on coverting GEOS-Chem diagnostic output from bpch to netCDF format.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>