The Species Database now tracks unique species names
In anticipation of FlexChem, we have modified the GEOS-Chem Species
Database object to track unique species names.  In other words, we
combine the list of advected tracers from "input.geos" and the list
of species in the KPP mechanism (from "gckpp_Monitor.F90) and remove
duplicate entries.

We also now store the indices of each species in the KPP "VAR" and
"FIX" arrays in the GEOS-Chem species database object.

This modification will allow us to reduce the amount of memory
needed for GEOS-Chem simulations (as we won't have to duplicate storage
for the same species in both State_Chm%TRACERS and State_Chm%SPECIES).

NOTE: For specialty simulations that do not use KPP, the list of unique
species names is just the list of advected tracers from input.geos.
Also, the KPP "VAR" and "FIX" indices will be set to missing values.

Also updated the Spc_Print routine in species_mod.F90 to print out
the extra KPP-related fields.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>