Remove code for initializing species concentrations from INIT_FLEXCHEM
In the netCDF restart file code, GEOS-Chem initializes species
concentrations to default values if the species data are not found in
the netCDF restart file. Therefore, the initial species concentrations
no longer need to be set in INIT_FLEXCHEM. We still convert the
units from mol/mol to molec/cm3 in INIT_FLEXCHEM if we're using the
background values (i.e. if SPC_IN_RST is false). The unit conversion is
kept in INIT_FLEXCHEM to reduce differences with pre-FlexChem code.
Moving the code to restart_mod.F introduces differences because of
slightly different temperature and pressure values used in CONV_FACTOR.

In addition, the call to READ_CSPEC_RESTART has been moved to
READ_GC_RESTART_BPCH in restart_mod.F to maintain the option of using
BPCH restart files. Code copied from READ_GC_RESTART_NC for setting default
background species concentrations if the species restart file is not found
or used has also been added to READ_GC_RESTART_BPCH. This is a temporary
fix because BPCH restart is slated for removal in the near future.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>