Now compile some KPP modules first to avoid circular dependencies
In the GEOS-Chem species database module, we need to have access to
the list of KPP species names so that we can combine the list of
advected tracers with the list of chemical species and remove duplicate
entries.  But the list of KPP species names is stored in KPP module
gckpp_Monitor.F90, which gets compiled after the species database
modules in Headers.

The solution is to add Makefile targets so that we can compile the
gckpp_Monitor.F90 (and its dependent modules gckpp_Precision.F90 and
gckpp_Parameters.F90) in a first pass, independently of the rest of
the KPP code.  This first pass is done immediately before the modules
in Headers are compiled, thus resolving the circular reference.

Also cleaned up some of the Makefiles and added extra comments.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>