Remove obsolete routines SETMODEL, RURALBOX, and AEROSOL_RURALBOX
In routine RDAER (in GeosCore/aerosol_mod.F),  the calculation of RELHUM
has been changed so that it doesn't depend on ABSHUM. We calculate RELHUM =
ABSHUM = AVGW / AIRNUMDEN, then calculate RELHUM = RELHUM / VPRESH2O. This
allows for the removal of State_Met%ABSHUM and also removal of the RURALBOX
and AEROSOL_RURALBOX routines that calculated that field. Those routines
also computed the number of chemistry boxes for use in 1D loops and arrays,
but previous commits have removed the necessity for those variables as
well. Ruralbox.F has been moved to the obsolete directory and the
AEROSOL_RURALBOX routine has been removed from aerosol_mod.F.

Setmodel.F has also been moved to the obsolete directory. It was
previously called from chemdr.F, but is no longer needed in the FlexChem
implementation.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>