Replace NLONG/ILONG, NLAT/ILAT, IPVERT with IIPAR, JJPAR, LLPAR where possible
In comode_loop_mod.F, we set several redundant variables for the number of
longitudes, latitudes, and levels. In summary:

   ILONG = NLONG = IIPAR
   ILAT  = NLAT  = JJPAR
   IVERT = IPVERT = NVERT = LLPAR (since IPLUME=0)

Here uses of NLONG/ILONG, NLAT/ILAT, and IPVERT have been removed where
possible and replaced with IIPAR, JJPAR, and LLPAR. This is a small step
towards the removal of comode_loop_mod.F. That module can't be removed
until we completely remove reader.F and readchem.F and we fully activate
the species database so that we can obtain values like NSPEC, NAMEGAS, etc.
from the species database.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>