Add species mapping vectors to State_Chm
We have added the mapping vectors:

  State_Chm%Map_Advect
  State_Chm%Map_Drydep
  State_Chm%Map_KppSpc
  State_Chm%Map_WetDep

which are intended to return the list of advected species, drydep
species, KPP species, and wetdep species from the State_Chm%SPECIES
field.  This will allow us to remove State_Chm%TRACERS completely,
or at least use it to point to the advected species in
State_Chm%SPECIES.

NOTE: For now the State_Chm%SPECIES array is still sized with the
IGAS parameter from SMVGEAR.  We have to leave it like this until
we completely remove all references to the SMVGEAR variables.

Also: in GeosCore/species_database_mod.F90, we now considered advected
species to be those listed in input.geos.  This is a more robust way
of flagging the advected species, especially since some species like
CO2 can be advected in one mechanism (e.g. standard) but non-advected in
another (e.g tropchem).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>