Now compute all drydep velocities explicitly
Our drydep scheme has the following identities:
(1) N2O5 uses the same drydep velocity as HNO3
(2) PMN, PPN, R4N2 use the same drydep velocity as PAN

In the prior code, we would assign the drydep ID of HNO3 to N2O5 and
the drydep ID of PAN to PMN, PPN, and R4N2 so that we could access
the drydep velocity from the proper slot of the DVEL array.  But this
algorithm meant that several species had the same drydep ID.  This
leads to ambiguity when trying to interface with the species database.

We have now modified the code so that a drydep velocity is explicitly
computed for N2O5, PMN, PPN, and R4N2.  This will ensure that these
species will have a unique drydep ID in the species database.  We made
the following modifications:

(1) Make sure that the DD_F0, DD_HStar_old, and MW_g values of HNO3
    get stored in the array slots for N2O5 in routine INIT_DRYDEP.

(2) Make sure that the DD_F0, DD_HStar_old, and MW_g values of PAN
    get stored in the array slots for PMN, PPN, and R4N2 in routine
    INIT_DRYDEP.

This update passed a geosfp_4x5_tropchem diff test on 19 May 2016.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>