Correct the units of the ND62 INST-MAP diagnostic
The default units of the State_Chm%TRACERS (aka STT) array were changed
from kg to kg/kg dry air in v11-01.  But the ND62 INST-MAP diagnostic
was not updated to take this into account.

When computing the column sums, we now multiply STT by the dry air
mass (AD) to convert from kg dry to kg.  Then the diagnostic will work
as expected.  We now have to pass State_Met as an argument to diag3.F
from main.F.

Thanks to Beata Bukosa (Wollongong) for pointing out this error!

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>