Bug fix: Make sure PEDGE is defined before calling Hco_CalcVertGrid
In routine GridEdge_Set (in module GeosCore/hcoi_gc_main_mod.F90), the
PEDGE variable was undefined, yet it was still passed to HEMCO routine
Hco_CalcVertGrid (in module HEMCO/Core/hco_geotools_mod.F90).  This
was probably causing the vertical grid for HEMCO to be defined
incorrectly.  But because there was no error trap to stop the run in
this case, we were unaware that this error situation persisted.

We have now made sure that PEDGE is properly defined and initialized
in GridEdge_Set before it is passed to Hco_CalcVertGrid.  We have also
added an error trap that will stop the GEOS-Chem simulation if
Hco_CalcVertGrid returns with an error.

NOTE: We know that much work remains to be done to bring the GC error
trapping up to date.  This work will need to be done after FlexChem is
implemented (and after SMVGEAR is removed).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>